requirements/local.txt
-r base.txt
-r mcp.txt

# Testing
pytest
pytest-django
pytest-mock
pytest-cov
factory-boy
